NBTify Upgrade - 1.90.0
Was having trouble upgrading this to the latest version, realized it was because Region-Types and Bedrock-LevelDB would be using different versions of NBTify, and that led to the type usage errors. So upgrading Region-Types to the latest version as well, seemed to fix the proglems I was having.

Going to start experimenting with implementing a tool to help transfer player inventories between worlds, thought it would only be of help to use the latest versions of my projects, since they have some features that I think I'm going to need here coming up soon. What's that? Oh, alright, gotcha.

https://stackoverflow.com/questions/1260748/how-do-i-remove-a-submodule
https://stackoverflow.com/questions/16773642/pull-git-submodules-after-cloning-project-from-github (refind, still not committed to memory yet)