Database Key Overload Usage
Realized how to fix this pattern/issue I've been having for a while for now. The concept is that the implementation function signature shouldn't use the extra-specific generics, but rather it should be available with an overload instead. So you get the internal type safety that comes with a regular function type definition, but you get the nice behavior from the outside level with the generic setup, allowing you to get more explicit type info on the use of the function itself! I'm not sure if this is a bad pattern to use two different definitions for this, it may fall out of sync with one another, but I'm going to ask around, and demo things out to see if this is a viable alternative to setting things up properly.