Mojang LevelDB Implementation!
Wow, what a day! Who'da thought that this would come about? Mojang released their own TypeScript implementation for LevelDB parsing, and it works in the browser! It seems to be for their own app, but it absolutely will work for anything else too, since it's just plain TS.

https://github.com/Mojang/minecraft-creator-tools/blob/b4748cdcc446edc85768eeab67c09848fc32b8c3/app/src/minecraft/LevelDb.ts

Big thanks to @SuperLlama88888 for letting me know about this!
https://github.com/SuperLlama88888/MCBE-LevelDB-reader

And of course big thanks to Mojang for making all of this possible to move forward with, to begin with!