node-leveldb-zlib Removal
Going to take the big step and see if this will work out accordingly. Hopefully it will! I could picture having memory issues for humungous worlds though. Might have to look into IndexedDB or some other way of doing things (in the browser; for Node I might have to build in my own file system for this implementation; at least we have the option for adding these in directly, now!).