ldb Key Parse Debug
Not working, I think the strings are probably encoded in a different order or something, I thought that this world was opening correctly before. This codebase intentionally errors currently, in the event of encountering a key that it doesn't know how to parse. I did this because I wanted to ensure I discovered as many keys as I could.